refactor: change relataed components folder to the app
