+++ src/app/films/components/FilmDetails/Skeleton.tsx
