+++ src/app/films/components/FilmDetails/index.tsx
